requirements.txt
PyQt6
opencv-python
PyQt6-WebEngine
requests
pyserial